+++ pyhadith/helpers.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """Helps in the pre-processing, analysis, post-processing and standardisation of ahadith."""
 
-import pyhadith.connector as connector
+from . import connector
 from nltk.stem.isri import ISRIStemmer
 # Import NTLK word_tokenize. If it fails, download 'punkt' and import again.
 try:
@@ -15,18 +15,15 @@ except:
 import csv
 import pyarabic.araby as araby
 
-def multReplace(text, tokens, rep):
-	""" Replaces a list of tokens ('tokens') within a string with another string ('rep')."""
-	for token in tokens:
-		text = text.replace(token, rep)
-	return text
-
-# This function cleans a text so that it may be processed later by the statistical models.
-def clean(text, words):
+# This function preprocesses a text so that it may be processed later by the statistical models.
+def preprocess(text, words):
 	"""Returns a string without punctuation, numbers, additional whitespace. Also adds whitespace before the arabic word 'wa' (and)."""
 	# A pre-defined list of tokens which must be removed from the text.
-	reps = ['.','/','<','>','?','؟','-','[',']','!',':','1','2','3','4','5','6','7','8','9','0','{','}','"',"'",'(',')',',','،','\n','\t']
-	text = multReplace(text, reps, ' ')
+	reps = ['.','/','<','>','?','؟','-','[',']','!',':','1','2','3','4','5','6','7','8','9','0','{','}','"',"'",'(',')',',','،','\n','\t','ـ','_','|','@','#','$','%','^','&','*','+','=','\\',
+	'~','`','a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z','A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P','Q','R','S',
+	'T','U','V','W','X','Y','Z']
+	for rep in reps:
+		text = text.replace(rep, ' ')
 	# Get rid of extra white space.
 	text = " ".join(text.split())
 	tokens = word_tokenize(text)
@@ -81,78 +78,145 @@ def isWa(token, words):
 	# Default to not 'wa'.
 	return False
 
+<<<<<<< HEAD
 def ahadith(text, arabicwords):
 	"""Uses the rawa and asl spaCy models to return a deconstructed and standardised hadith object."""
+=======
+def deconstruct(text):
+	"""Deconstructs a hadith into a matn and an isnad, using the 'ajza', 'musaid' and 'rawa' models.
+	Returns matn and isnad objects."""
+>>>>>>> v0.1.0
 
-	doc = connector.process(text, 'rawa')
-	# Init the hadith object.
-	hadith = {
-		"isnad" : {
-			"raw" : text, # Default to isnad being the entire text.
+	isnad = {
+			"raw" : text, # Defaults to entire text.
 			"start_char" : 0,
 			"end_char" : len(text)-1,
 			"narrators" : []
-		},
-		"matn" : {
+		}
+	
+	matn = {
 			"raw" : "",
 			"start_char" : None,
 			"end_char" : None,
-		},
-		"category" : {"name" :"", "score" : ""},
-	}
-
-	catsDoc = connector.process(text, 'asl')
-	# Set the category to be athar if the athar score is greater than the khabar score, or else default to khabar.
-	if catsDoc.cats['athar'] > catsDoc.cats['khabar']:
-		hadith['category']['name'] = 'athar'
-		hadith['category']['score'] = catsDoc.cats['athar']
-	else:
-		hadith['category']['name'] = 'khabar'
-		hadith['category']['score'] = catsDoc.cats['khabar']
+		}
 	
-	# Segment the hadith into an isnad at the last occurance of a narrator's name.
-	lastIsnadEnd = len(text)-1
-	for ent in doc.ents:
-		lastIsnadEnd = ent.end_char
-	if (lastIsnadEnd != len(text)-1):
-		hadith['isnad']['raw'] = text[:lastIsnadEnd]
-		hadith['isnad']['end_char'] = lastIsnadEnd
-		hadith['matn']['raw'] = text[lastIsnadEnd:]
-		hadith['matn']['start_char'] = lastIsnadEnd
-		hadith['matn']['end_char'] = len(text)-1
+	ajzaDoc = connector.process(text, 'ajza')
+	musaidDoc = connector.process(text, 'musaid')
 
-	# If isnad is empty, return the hadith, or else begin processing the narrators in the isnad.
-	if hadith['isnad']['end_char'] == 0:
-		return hadith
+	# Split the hadith into a 'isnad' and 'matn' at the word succeeding the last narrator preceding the last 'STARTMATN' tag.
+
+	# Look for the last 'STARTMATN' tag.
+	tokenCounter = 0
+	ajzaBreak = None
 	
-	ents = []
+	for token in ajzaDoc:
+		tokenCounter = tokenCounter+1
+		if token.tag_ == 'STARTMATN':
+			ajzaBreak = tokenCounter
+	
+	if ajzaBreak == None:
+		ajzaBreak = tokenCounter
+	
+	lastToken = tokenCounter
+	
+	# Look for the last narrator before the last 'STARTMATN' tag.
+	tokenCounter = 0
+	musaidBreak = None
+	musaidBreakChar = None
+	tokensWithoutNarrator = 0
+	for token in musaidDoc:
+		tokenCounter = tokenCounter+1
+		if tokenCounter <= ajzaBreak:
+			# Deal only with tokens that are classed as RAWINAME and are not '\u200f'.
+			if token.ent_type_ == 'RAWINAME' and token.text != '\u200f':
+				musaidBreak = tokenCounter+1
+				tokensWithoutNarrator = 0
+			else:
+				tokensWithoutNarrator = tokensWithoutNarrator+1
+		else:
+			break
+	
+	# Store the 'isnad' and 'matn' data.
+	tokenCounter = 0
+	lastChar = 0
+	if musaidBreak:
+		isnad['raw'] = []
+		matn['raw'] = []
+		for token in musaidDoc:
+			tokenCounter = tokenCounter+1
+			if tokenCounter > musaidBreak:
+				matn['raw'].append(token.text)
+			else:
+				isnad['raw'].append(token.text)
+		isnad['raw'] = " ".join(isnad['raw'])
+		isnad['start_char'] = 0
+		isnad['end_char'] = len(isnad['raw'])
+		matn['raw'] = " ".join(matn['raw'])
+		matn['star_char'] = isnad['end_char']
+		matn['end_char'] = len(text)
+
+	# Send the 'isnad' text to rawa.
+	rawaDoc = connector.process(isnad['raw'], 'rawa')
+
 	# Set a closed class of stemmed join terms.
 	joinTerms = ['حدث','عن','قال','ثنا','خبر','نا','وقل','نبأ','ان','انا', 'قلا','انه','سمع','أخبر','يقل','وقل','غدد','قرء','أصب',
 	'قلت','مرو','عنى','بمك','وعن','في','بمك','كوف','ملء','ذكر','من','لفظ','او','وهذا','هذا', 'كلهم','تقل','وكانت','كان']
-	for ent in doc.ents:
+	# Set a closed class of excluded narrator names.
+	narratorNames = ['رضى الله عنهما','رضى','الله','عنهما','رضى الله عنهم','\u200f']
+	# Add entities found by spaCy to isnad['narrators'].
+	for ent in rawaDoc.ents:
 		entText = ent.text
 		label_ = ent.label_
 		start_char = ent.start_char
 		end_char = ent.end_char
-		dont = 0
+		add = True
+		# Deal with empty narrators.
+		if " ".join(word_tokenize(entText)).replace(' ','').replace('\u200f','') != '':
+			# Deal with cases where first word of ent is a join term.
+			stemmer = ISRIStemmer()
+			words = (" ".join(entText.split())).split(" ")
+			firstWord = words[0]
+			firstStem = stemmer.stem(araby.strip_tashkeel(firstWord))
+			if firstWord in joinTerms:
+				# Don't add this ent if its in the join terms AND its only one word.
+				if len(words) == 1:
+					add = False
+				else:
+					del words[0]
+					entText = " ".join(words)
+					start_char = start_char+len(words)
+			# Deal with cases where the narrator name is an excluded name.
+			for excludedName in narratorNames:
+				if entText.replace(excludedName,'').replace(' ','') == '':
+					add = False
+			# Add to isnad['narrators'] if there is a still name after having removed join terms.
+			if add == True:
+				isnad['narrators'].append({'text' : entText, 'label_' : label_,'start_char' : start_char, 'end_char' : end_char})
 
-		# Deal with cases where first word of ent is a join term.
-		stemmer = ISRIStemmer()
-		words = (" ".join(entText.split())).split(" ")
-		firstWord = words[0]
-		firstStem = stemmer.stem(araby.strip_tashkeel(firstWord))
-		if firstWord in joinTerms:
-			# Don't add this ent if its in the join terms AND its only one word.
-			if len(words) == 1:
-				dont = 1
-			else:
-				del words[0]
-				entText = " ".join(words)
-				start_char = start_char+len(words)
-				end_char = end_char+len(end_char)
-		# Add to ents if we're allowed to.
-		if dont == 0:
-			ents.append({'text' : entText, 'label_' : label_,'start_char' : start_char, 'end_char' : end_char})
+	return isnad, matn
+
+def categorize(text):
+	"""Uses the 'asl' model to categorize a hadith as either an atar or a khabar.
+	Returns the label of the category and the score assigned to the categorization by the 'asl' model."""
+
+	doc = connector.process(text, 'asl')
+
+	# Set the category to be atar if the atar score is greater than the khabar score, or else default to khabar.
+	if doc.cats['atar'] > doc.cats['khabar']:
+		name = 'atar'
+		score = doc.cats['atar']
+	else:
+		name = 'khabar'
+		score = doc.cats['khabar']
+	
+	return {"name" : name, "score" : score}
+	
+def treeify(isnad, words):
+	"""Reconstructs the isnad of a given hadith.
+	Returns a tree-like data structure representing narrational relationships of narrators in the isnad."""
+	tree = []
+	ents = isnad['narrators']
+	text = isnad['raw']
 	# Create a In-Out list for tokens in the isnad.
 	narrators = []
 	IOisnad = []
@@ -164,7 +228,7 @@ def ahadith(text, arabicwords):
 		start_char = ent['start_char']
 		end_char = ent['end_char']
 		# If ent is out of isnad, break the loop.
-		if end_char > hadith['isnad']['end_char']:
+		if end_char > isnad['end_char']:
 			break
 		# If ent isn't the first token, add the text before it.
 		if start_char != 0:
@@ -193,8 +257,13 @@ def ahadith(text, arabicwords):
 			if narratorCount == 1:
 				parents = None
 			# If prev. item is OUT and is wa, assume parents of previous narrator.
+<<<<<<< HEAD
 			elif IOisnad[count-2]['label'] == 'OUT' and isWa(" ".join(IOisnad[count-2]['raw'].split()).split(' ')[0], arabicwords):
 				parents = hadith['isnad']['narrators'][narratorCount-2]['parents']
+=======
+			elif IOisnad[count-2]['label'] == 'OUT' and isWa(" ".join(IOisnad[count-2]['raw'].split()).split(' ')[0], words):
+				parents = tree[narratorCount-2]['parents']
+>>>>>>> v0.1.0
 			# Else, look for parents.
 			else:
 				parents = []
@@ -224,7 +293,7 @@ def ahadith(text, arabicwords):
 					parents.append(narratorCount-pCount)
 				
 			# Append to narrators
-			hadith['isnad']['narrators'].append(
+			tree.append(
 				{
 					"id" : narratorCount,
 					"name" : item['raw'],
@@ -233,4 +302,4 @@ def ahadith(text, arabicwords):
 					"end_char" : item['end_char']
 				}
 			)
-	return hadith
+	return tree